fixed pytest tests

+++ tests/conftest.py
@@ -1,4 +1,5 @@
 import pytest
+import os
 from mdtemplate import create_form
 from shutil import rmtree
 
@@ -6,7 +7,9 @@ from shutil import rmtree
 @pytest.fixture()
 def remove_temporary_dir_after_test():
     yield
-    rmtree(create_form.TEMP_DIR)
+    if os.path.exists(create_form.TEMP_DIR):
+        rmtree(create_form.TEMP_DIR)
+
 
 @pytest.fixture()
 def reset_temp_dir():
